Painonsyöttökenttää paranneltu, käyttäjätaulua alustettu tietokantaan

+++ loppuprojekti_spring/src/main/java/fi/academy/paino/Paino.java
@@ -19,11 +19,11 @@ public class Paino {
     private String kayttajaId;
     // ei ole aivan varmaa vielä, onko LocalDate sopiva formaatti
     private LocalDate pvm;
-    private Integer painoKiloina;
+    private Double painoKiloina;
 
     public Paino() { }
 
-    public Paino(String kayttajaId, LocalDate pvm, int paino) {
+    public Paino(String kayttajaId, LocalDate pvm, double paino) {
         this.kayttajaId = kayttajaId;
         this.pvm = pvm;
         this.painoKiloina = paino;
@@ -62,11 +62,11 @@ public class Paino {
         this.pvm = pvm;
     }
 
-    public Integer getPainoKiloina() {
+    public Double getPainoKiloina() {
         return painoKiloina;
     }
 
-    public void setPainoKiloina(Integer paino) {
+    public void setPainoKiloina(Double paino) {
         this.painoKiloina = paino;
     }
 }
